refactor(plugins,ui): split the remaining oversized GitHub-sync test files
Three files crossed the 300-line cap (finding #13):

- store_tests.rs (367 lines) -> store_tests/{migrations,link,pairs,runs}.rs,
  mirroring the table groups store.rs itself exposes.
- lib/api/__tests__/todos-github.test.ts (444 lines, grew past the cap once
  workflowLabels landed) -> todos-github-{link,pairs,sync}.test.ts, one per
  client surface.
- features/tasks/github/__tests__/use-github-sync-store.test.ts (494 lines)
  -> use-github-sync-store-{load,mutations,dialog}.test.ts, one per store
  behavior group.

Each split file stays self-contained (duplicated fixtures/mocks, no shared
support module), matching the existing convention in
lib/api/__tests__/http-*.test.ts. No behavior changes.
